Add 'command' to chat telemetry (#219218)
