+++ src/vs/workbench/contrib/inlineChat/browser/inlineChatController.ts
@@ -1074,6 +1074,7 @@ export class InlineChatController implements IEditorContribution {
 				sessionId: response.session.sessionId,
 				requestId: response.requestId,
 				agentId: response.agent?.id,
+				command: response.slashCommand?.name,
 				result: response.result,
 				action: {
 					kind: 'inlineChat',
@@ -1099,6 +1100,7 @@ export class InlineChatController implements IEditorContribution {
 				sessionId: response.session.sessionId,
 				requestId: response.requestId,
 				agentId: response.agent?.id,
+				command: response.slashCommand?.name,
 				result: response.result,
 				action: {
 					kind: 'inlineChat',
